refactor: expose types centrally and use that
